+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to `phpcs-type-sniff` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Next
-## Changed
+### Removed
+- Setting `FqcnMethodSniff.usefulTags`
+
+### Added
+- Setting `FqcnMethodSniff.invalidTags`. Specifies which methods tags should be removed.
+
+### Changed
 - Parsing of generic tags with parentheses, no longer part of tag name, Dynamic content is now tag content.
+- All tags except `param`, `return` make method PHPDoc useful.
 
 ## 0.12.4 - 2019-11-21
 ### Fixed